requirements.txt
flask
requests
firebase-admin
google-generativeai
